renamed entry file to app.ts
