+++ .vscode/launch.json
@@ -5,7 +5,7 @@
             "name": "Launch",
             "type": "node",
             "request": "launch",
-            "program": "${workspaceRoot}/src/sikky.ts",
+            "program": "${workspaceRoot}/src/app.ts",
             "stopOnEntry": false,
             "args": [],
             "cwd": "${workspaceRoot}",
