+++ package.json
@@ -2,8 +2,8 @@
   "name": "sikky",
   "version": "1.1.4",
   "description": "A ultimate TypeScript 2.0 boilerplate for node and the browser",
-  "main": "dist/sikky.js",
-  "module": "dist/finale/sikky.mjs",
+  "main": "dist/app.js",
+  "module": "dist/finale/app.mjs",
   "modules.root": "build/src",
   "author": "KFlash",
   "homepage": "https://github.com/kflash/sikky",
